+++ frontend/src/views/FieldWellKnown.svelte
@@ -2,6 +2,7 @@
   import FieldText from "./FieldText.svelte";
   import FieldBoolValue from "./FieldBoolValue.svelte";
   import FieldNilText from "./FieldNilText.svelte";
+  import FieldTimestamp from "./FieldTimestamp.svelte";
 
   export let name = "";
   export let message = {};
@@ -11,15 +12,13 @@
 
   let field = {};
   let placeholder = "";
+
   $: {
     field = {
       name: name,
       kind: message.full_name,
     }
 
-    if (field.kind === "google.protobuf.Timestamp") {
-      placeholder = "2006-01-02T15:04:05.000Z";
-    }
     if (field.kind === "google.protobuf.Duration") {
       placeholder = "0.1s";
     }
@@ -40,6 +39,10 @@
 {:else if field.kind === "google.protobuf.BytesValue"}
 
   <FieldNilText on:remove {field} {state} {key} {idx} multiline />
+
+{:else if field.kind === "google.protobuf.Timestamp"}
+
+<FieldTimestamp on:remove {field} {state} {key} {idx} />
 
 {:else}