+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Date picker for `google.protobuf.Timestamp` fields
 - Request type icons, in-flight indicator and network out/in count
 
 ### Fixed